Dump of some *very* in-progress prototype code.
Consider this code speculative. Largely it is experimentation with the APIs that will be restructured into a more refined product.
Presently installed Apps can be found and used to generate a rough F-Droid index.xml.
Separately, an embedded NanoHTTPD engine is set up, ready to be configured to serve the index.xml & APKs.