Fix the APK <sig> element calculation.
Previously I was taking the SHA1 of the bytes[] of the entire opaque PKCS7 object that Android uses as the apk 'Signature'.

The f-droid server seems to instead take the java Certificate from the PKCS7 object, convert its bytes into hex and then take the MD5 of the hex string as bytes. I've yanked the calculation code directly from the fdroidserver because I'm not 100% confident I understnad how/why they are doing this conversion and at this point I just want compatibility.

Moving forward there are much better ways to be doing this.