store the IP Address and repo URL at the start so we can gen QRCodes, etc.
The IP Address is used as the repo address, which is also used for the
direct download link to fdroid.apk for bootstrapping other devices. It is
split out into its own method so that it can be called whenever the IP
address changes, i.e. when the device connects to a new wifi network.

refs #2454