if wifi is not enabled, make the repoSwitch enable wifi, then the repo
If the wifi is not enabled, then there is no IP address to get, and
therefore the repo cannot be configured.  To handle this, the repoSwitch
is turned into a switch for enabling the wifi first.  Once the wifi is
enabled, then the repoSwitch is configured to start/stop the repo.  The
repo URL is displayed in the repoSwitch, and the wifi network name is
displayed in its own text field below it.

refs #2454