Sign the index.jar file using Kerplapp keypair/signing cert.
No longer using the testkey mode of zipsigner-lib to sign the index.jar.
Signature is now applied using the keypair & cert associated with the
INDEX_CERT_ALIAS in the keystore.

Without zipsigner-lib-optional on the classpath zipsigner generates
ClassNotFound exceptions when using custom keys. Unfortunately this means yet
another jar in the libs directory :-/