Generate an index.html to bootstrap the repo.
Implements Feature #2668 : index.html for bootstrap.

When the IP address is set from wifi the KerplappRepo now generates an
index.html in the wwwroot of the repo. If the Kerplapp host has the F-droid
client APK installed it will be symlinked (or copied) into the wwwroot and used
by the bootstrap page for clients to install F-droid. In the event you have just
Kerplapp and not F-droid the link points to the internet accessible F-droid APK.