Added a Subject Alternative Name w/ local IP to https cert.
The default StrictHostnameVerifier used by the SSL stack on Android/Java does
not accept certificates with the CN field equal to an IP addres, complaining
about hostname validation failing. This happens regardless of the TrustManager
implementation, so even using MemorizingTrustManager or
AndroidPinningTrustManager does not solve the problem. Adding a X509 extension
specifying the IP address as a Subject Alternative name of type IP is enough to
make the default StrictHostnameVerifier implementation happy. This avoids us
having to hack about with an overloaded implementation that may be unsafe. As
good a compromise as any!