Remember local repo setup for App select activity.
Kerplapp should pre-select applications on the repo app select activity
that have been previously added to the repo. This is accomplished by
modifying the AppListLoader to set the loaded AppEntry to 'enabled' when
there is a corresponding APK file in the repo directory from a previous
setup.

This patch addresses bug #2469 - 'remember local repo setup'.