Adding rudimentary Network Service Discovery.
Service advertisement is wired up to the HTTPD instance & wifi status.
Presently using the built-in Android support for NSD added since API
level 16 (Jelly-Bean). This will need to be addressed if we want to keep
the Kerplapp target-api at a reasonable level. Likely we will have to
translate the NsdHelper to use jmDNS or another external lib we can
ship.