add support for sending repo URI via NFC
If you are on the Kerplapp main screen and you do the NFC bump, and touch
Kerplapp to send, the URI of Kerplapp's repo will be sent via NFC to the
other side.

Because NFC callbacks must be attached to an Activity, and the relevant
Interfaces weren't added until android-14, this includes a hack to launch a
different Activity depending on the platform that the app is currently
running on.  It seems to be the only way...