fix crash when reading maxage pref
an EditTextPreference will always return a String, so getInt() causes a
ClassCastException.  The EditTextPreference for maxage needs to be
retreived as a String, then converted.  It is first converted to a Float,
in case the user has entered in a decimal, and that is then forced to an
int value.