run icons copy in background AsyncTask, its optional
The user does not need to wait for the icons copy to finish in order to use
the new repo, it will work fine without the icons.  So put the icons copy
in an AsyncTask without a progress dialog so that it happens in the
background while the user can do other things.

refs #2694