Use an FDroid specific NSD service type, based on SSL preference.
NSD service types follow a DNS SRV convention _protocol._transport used in
mDNS. Previously Kerplapp was advertising an _http._tcp service
generically. It makes more sense to advertise a _fdroidrepo._tcp service so
the FDroid client can run a discovery that it knows won't return generic
non-FDroid repo HTTP servers.

Further, we should advertise _fdroidrepo._tcp when the use_ssl preference is
disabled and _fdroidrepos._tcp when it is enabled to indicate to the
discovering peer whether we are advertising an HTTP or an HTTPS FDroid repo.

refs #2467