set default repo name to the BRAND/MODEL of the device
Having the default repo name be the BRAND/MODEL should make it easier to
identify the device in question without broadcasting sensitive info like
unique IDs, names, or email addresses.  Users can still easily customize
the name in the Settings.

Now that we have a more descriptive repoName, use it anywhere it makes
sense to have it.