Fix crash on Samsung device from missing PackageInfo.
It seems on some devices there can be a ApplicationInfo object, but we
can't get an AppInfo object for the corresponding package from
PackageManager.getPackageInfo(). This was causing a NPE where we assumed
the AppInfo object would never be null.

The fix is skipping apps that don't have an AppInfo. We need this
structure to be able to learn the version code and other necessary
information for the repo index.

I believe this fixes Issue #3020.